Add support for Indirect Index
